Rename database/{gravity.* => gravity-db.*}
